A basic program structure to start from. Essentially loads the Sig Processing module page and prints the urls of all pdf files. Requires UoY credentials.
Pretty simple right?!
There's some extra bits in there that allow to operate in debug mode. Essentially how this works is that it downloads some reference files that can be used for offline parsing so we're not constantly hitting the server. Didn't want to include the ref html files on the repo because they contain some personal info like emails of lecturers - so the program generates them on first run.